pyproject.toml
[project]
name = "align-tune"
description = "AlignTune: Enhancing Image Captioning via Text Data Augmentation."
authors = [{"name" = "Umut DUNDAR", "email" = "dundar.umut@metu.edu.tr"}]
license = "MIT"
dynamic = ["version", "readme"]

requires-python = ">=3.10.0, <3.11.0"

dependencies = [
"hydra-core",
"lightning >= 2.2.1",
"wandb >= 0.16.4",
"pandas >= 2.2.1",
"numpy >= 2.1.3",
"typer >= 0.4.0",
"pillow >= 10.0.0",
"torchvision >= 0.16.1",
"albumentations >= 1.4.22",
"typer >= 0.4.0",
"pillow >= 9.0.0",
"tqdm >= 4.64.0",
"safetensors >= 0.3.0",
"nltk >= 3.8.1",
"peft >= 0.4.0",
"wordcloud >= 1.9.4",
"pycocoevalcap >= 1.2",
"bitsandbytes == 0.45.5",
"seaborn",
"matplotlib",
"transformers == 4.51.3",
]


[project.optional-dependencies]
dev = ["pre-commit", "mypy", "setuptools", "wheel", "build", "pytest"]

# [project.scripts]
# align_tune = "align_tune.__main__"

[tool.setuptools.dynamic]
readme = { file = "README.md" }
version = { attr = "aligntune.__version__" }

[tool.setuptools.packages.find]
exclude = ["data"]

[build-system]
requires = ["wheel", "setuptools"]
build-backend = "setuptools.build_meta"
